+++ aries-site/src/data/structures/extend.js
@@ -1,0 +1,160 @@
+/* eslint-disable max-len */
+export const extend = [
+  {
+    name: 'API Chomp',
+    available: false,
+    description:
+      'Got API, but need to easily pull data out of it for use in your app or designs. Chomp! Chomp!',
+    preview: {
+      image: {
+        src: '/extendImages/extend-preview-chomp.svg',
+        alt: 'HPE API Chomp Preview',
+      },
+    },
+    seoDescription:
+      'Got API, but need to easily pull data out of it for use in your app or designs. Chomp! Chomp!',
+    sections: [],
+  },
+  {
+    name: 'Designer',
+    available: true,
+    description:
+      'Build HPE compliant designs in code that are ready to be consumed on the web.',
+    preview: {
+      image: {
+        src: '/extendImages/extend-preview-designer.svg',
+        alt: 'HPE Designer Preview',
+      },
+    },
+    seoDescription:
+      'Build HPE compliant designs in code that are ready to be consumed on the web.',
+    sections: [],
+  },
+  {
+    name: 'Designer Guidance',
+    available: true,
+    description:
+      'Starter files, patterns, interactions, and workflows on how to succeed using the design resources from HPE Design System and the HPE Brand.',
+    seoDescription:
+      'View patterns, interactions, and other best practices for how to succeed using resources included with the HPE Design System.',
+    sections: [
+      'Getting started',
+      'Setting up your Figma account',
+      'Joining the HPE Design System Figma team',
+      'Before you start desiging',
+      'HPE Design System Figma Library',
+    ],
+    relatedContent: ['Components', 'Templates', 'Designer'],
+  },
+  {
+    name: 'Developer Guidance',
+    available: true,
+    description:
+      'Resources for setting up your application with the HPE Design System library and HPE theme.',
+    seoDescription:
+      'Resources for setting up your application with the HPE Design System library and HPE theme.',
+    sections: [
+      'Getting started',
+      'Preferred environment',
+      'ReactJS and Grommet starter resources',
+      'Applying the HPE theme',
+      "What if our team doesn't use ReactJS?",
+    ],
+    relatedContent: ['Components', 'Templates', 'Designer'],
+  },
+  {
+    name: 'Global Sidebar',
+    available: false,
+    description:
+      'For use when building applications and services that live in the HPE eccosystem.',
+    preview: {
+      image: {
+        src: {
+          light: '/extendImages/extend-preview-sidebar.svg',
+          dark: '/extendImages/extend-preview-sidebar-invert.svg',
+        },
+        alt: 'HPE Global Sidebar Preview',
+      },
+    },
+    seoDescription:
+      'For use when building applications and services that live in the HPE eccosystem.',
+    sections: [],
+  },
+  {
+    name: 'HPE Audience',
+    available: false,
+    description:
+      'Keep it simple! Present and share slides with this lightweight service.',
+    preview: {
+      image: {
+        src: {
+          light: '/extendImages/extend-preview-slides-audience.svg',
+          dark: '/extendImages/extend-preview-slides-audience-invert.svg',
+        },
+        alt: 'HPE Audience Preview',
+      },
+    },
+    seoDescription:
+      'Keep it simple! Present and share slides with this lightweight service.',
+    sections: [],
+  },
+  {
+    name: 'HPE Docs',
+    available: false,
+    description:
+      'The only CMS you will ever need! Publish content quickly and easily.',
+    preview: {
+      image: {
+        src: {
+          light: '/extendImages/extend-preview-hpe-docs.svg',
+          dark: '/extendImages/extend-preview-hpe-docs-invert.svg',
+        },
+        alt: 'HPE docs Preview',
+      },
+    },
+    seoDescription:
+      'The only CMS you will ever need! Publish content quickly and easily.',
+    sections: [],
+  },
+  {
+    name: 'HPE Images',
+    available: false,
+    description:
+      'Need a place to host images in the cloud for use by your HPE services and apps?',
+    preview: {
+      image: {
+        src: '/extendImages/extend-preview-hpe-images.png',
+        alt: 'HPE Images Preview',
+      },
+    },
+    seoDescription:
+      'Need a place to host images in the cloud for use by your HPE services and apps?',
+    sections: [],
+  },
+  {
+    name: 'React Router',
+    available: false,
+    description:
+      'A collection of navigational components that compose declaratively with your application.',
+    seoDescription:
+      'A collection of navigational components that compose declaratively with your application.',
+  },
+  {
+    name: 'Table Topper',
+    available: false,
+    description:
+      'Build tables for your app quickly with all the filtering, sorting, and controls you could need!',
+    preview: {
+      image: {
+        src: {
+          light: '/extendImages/extend-preview-table.svg',
+          dark: '/extendImages/extend-preview-table-invert.svg',
+        },
+        alt: 'HPE Table Topper Preview',
+      },
+    },
+    seoDescription:
+      'Build tables for you app quickly with all the filtering, sorting, and controls you could need!',
+    sections: [],
+  },
+];
