+++ aries-site/src/data/structures/index.js
@@ -1,0 +1,5 @@
+export * from './components';
+export * from './extend';
+export * from './foundation';
+export * from './templates';
+export * from './Structure';
